pyproject.toml
[tool.poetry]
name = "fastapi-backend-with-unit-integration-tests"
version = "0.1.0"
description = ""
authors = ["nickTheof <156532831+nickTheof@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.2"}
sqlalchemy = "^2.0.32"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
pyjwt = "^2.9.0"
passlib = "^1.7.4"
bcrypt = "4.0.1"
black = "^24.8.0"
alembic = "^1.13.2"
httpx = "^0.27.2"
jinja2 = "^3.1.4"
aiofiles = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
